LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigatesuccess-same-document-expected.txt


FAIL navigatesuccess fires for a same-document navigation assert_equals: expected "#1" but got ""
